+++ ruby_llm_eval/assets/tasks/004_flatten_nested_array/solution_ref.rb
@@ -1,0 +1,9 @@
+def flatten_deep(arr)
+  arr.each_with_object([]) do |element, out|
+    if element.is_a?(Array)
+      out.concat(flatten_deep(element))
+    else
+      out << element
+    end
+  end
+end